Plants now work properly
Found that there was an error occurring with checking for different
heights. The `possible.location` matrix was not a matrix and was causing
an error in the `for(i in 1:nrow(possible.location))` loop. To fix this
problem I added `if(class(possible.location) != 'matrix')` with the same
code in the `for` loop copied into it so that it would still check for
positions outside of the plants "habitat".