Increased readability of 'tester.R' and other small changes
I added more comments througout the code to further explain what is
happening inside the given function (or code chunk depending on size).
I change the 'new.plant.loc' function again to return the propper
value. Minor changes have been made within the herbivore functions
so values are updated and returned propperly.

This commit is occuring before I begin work on the 'eat' function
for herbivores. This function will have to have a check inside of it
that takes the bases of the herbivores and converts them into enzymes.
This will actually be done in another function and called in the 'eat'
function.